Fixes for the PR
