+++ src/common/navbar/navbar.tsx
@@ -1,6 +1,7 @@
 import { AppLink, RouteName } from 'routes';
 import { AppContext } from 'context';
 import { useContext } from 'react';
+import { classnames } from 'helpers/utils';
 
 import styles from './navbar.module.scss';
 import { ReactComponent as ProfileSVG } from '../../assets/icons/profile.svg';
@@ -12,11 +13,11 @@ export const Navbar = () => {
   return (
     <div className={styles.navbar}>
       <AppLink routeName={RouteName.Home}>
-        <HomeSVG className={`${styles.homeIcon} ${styles.singleIcon}`} />
+        <HomeSVG className={classnames(styles.homeIcon, styles.singleIcon)} />
       </AppLink>
       <div className={styles.profileDiv}>
         {`Hola, ${state.firstName}`}
-        <ProfileSVG className={`${styles.profileIcon} ${styles.singleIcon}`} />
+        <ProfileSVG className={classnames(styles.profileIcon, styles.singleIcon)} />
       </div>
     </div>
   );
