+++ src/common/sign-in/sign-in.tsx
@@ -4,9 +4,9 @@ import { TextField, TextFieldStatus } from 'common/text-field';
 import { ToastContainer, toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 import { useNavigate } from 'react-router-dom';
-import { SignInController } from 'networking/controllers/sign-in-controller';
+import { UserController } from 'networking/controllers/user-controller';
 import { userActions, AppContext } from 'context';
-import { UserSerializer } from 'networking/serializers/users-serializer';
+import { UserSerializer } from 'networking/serializers/user-serializer';
 
 import { ReactComponent as EyeSVG } from '../../assets/icons/eye.svg';
 import { ReactComponent as ClosedEyeSVG } from '../../assets/icons/closed-eye.svg';
@@ -40,14 +40,16 @@ export const SignIn = () => {
 
   const loginUser = async () => {
     try {
-      const userSignIn: SignIn = {
+      const userSignIn: User = {
         email: emailState.inputValueEmail,
         password: passState,
+        firstName: '',
+        lastName: '',
+        token: '',
       };
 
       setBtnIsLoading(true);
-      await SignInController.signIn(userSignIn);
-      const { data } = await SignInController.signIn(userSignIn);
+      const { data } = await UserController.signIn(userSignIn);
       dispatch({ type: userActions.USER_LOGGED, user: UserSerializer.deSerialize(data) });
       notifyOk();
