--- src/networking/serializers/sign-in-serializer.ts
@@ -1,17 +1,0 @@
-class SignInSerializer {
-  static deSerialize(data: RawSignIn): SignIn {
-    return {
-      email: data.email,
-      password: data.password,
-    };
-  }
-
-  static serialize(data: SignIn): RawSignIn {
-    return {
-      email: data.email,
-      password: data.password,
-    };
-  }
-}
-
-export { SignInSerializer };
