--- src/networking/types/sign-in.d.ts
@@ -1,9 +1,0 @@
-type SignIn = {
-  email: string,
-  password: string,
-};
-
-type RawSignIn = {
-  email: string,
-  password: string,
-};
