+++ src/networking/types/user.d.ts
@@ -1,13 +1,15 @@
-type SignUp = {
+type User = {
   firstName: string,
   lastName: string,
   email: string,
+  token: string,
   password: string,
 };
 
-type RawSignUp = {
+type RawUser = {
   firstname: string,
   lastname: string,
   email: string,
+  token: string,
   password: string,
 };
