--- src/networking/types/users.d.ts
@@ -1,13 +1,0 @@
-type User = {
-  firstName: string
-  lastName: string
-  email: string
-  token: string
-};
-
-type RawUser = {
-  firstname: string
-  lastname: string
-  email: string
-  token: string
-};
